Move rollPrompt to appropriate place in code
rollPrompt and its types did not belong in the data folder
